+++ api/tests/certification/enrolment/integration/infrastructure/repositories/session-repository_test.js
@@ -1,6 +1,6 @@
-import { NotFoundError } from '../../../../../../lib/domain/errors.js';
 import { SessionEnrolment } from '../../../../../../src/certification/enrolment/domain/models/SessionEnrolment.js';
 import * as sessionRepository from '../../../../../../src/certification/enrolment/infrastructure/repositories/session-repository.js';
+import { NotFoundError } from '../../../../../../src/shared/domain/errors.js';
 import { catchErr, databaseBuilder, domainBuilder, expect, knex } from '../../../../../test-helper.js';
 
 describe('Integration | Repository | certification | enrolment | SessionEnrolment', function () {
@@ -100,6 +100,34 @@ describe('Integration | Repository | certification | enrolment | SessionEnrolmen
     it('should return a Not found error when no session was found', async function () {
       // when
       const error = await catchErr(sessionRepository.get)({ id: 2 });
+
+      // then
+      expect(error).to.be.instanceOf(NotFoundError);
+    });
+  });
+
+  describe('#getVersion', function () {
+    let session;
+
+    beforeEach(async function () {
+      // given
+      session = databaseBuilder.factory.buildSession({
+        version: 3,
+      });
+      await databaseBuilder.commit();
+    });
+
+    it('should the version of the session if it exists', async function () {
+      // when
+      const version = await sessionRepository.getVersion({ id: session.id });
+
+      // then
+      expect(version).to.equal(3);
+    });
+
+    it('should return a Not found error when no session was found', async function () {
+      // when
+      const error = await catchErr(sessionRepository.getVersion)({ id: 2 });
 
       // then
       expect(error).to.be.instanceOf(NotFoundError);
